Add query connection for comparison
